DynamicStreamSet is now specified to remove streams which complete
Added a test that ensures streams which poll to completion are removed
from a DynamicStreamSet automatically.

I had to add handles to do this, but they now allow editing a
DynamicStreamSet while something else polls it. Only the holder of the
set itself can update it, but all handles can poll it. Only one handle
will receive each item that arrives, so multiple handles will distribute
their outputs, and not necessarily evenly.