Tunnel unification refactoring
Now Tunnel is a hybrid trait of uplink and downlink. Uplink and Downlink
are traits, and sidedness is split off into its own trait. The hybrid
tunnel allows any number of accesses to the incoming stream, but they
are exclusive, and will await an async Mutex- possibly until the tunnel
in question has closed.

Future revisions may include an API to split tunnels, and will
include control operations and event monitoring to allow connections
to be closed on command, or observed for current and future state.