Crate `fred` lacks `mocks` feature; crates.io disallows wildcards
This was added to enable the feature only during dev-dependencies, but
it is unnecessary and a no-op. While it remains unanswered how to tweak
a dev-dependency without altering the version constraint, this has been
removed to allow publishing.

See https://github.com/aembke/fred.rs/issues/36 for details about the
mocks layer being absent from current versions of the `fred` crate.